IDSR/DMI Flags optimization - removed 8 flags: combined them into IDSR flags for performance
